added login

+++ Lingoleap - BE/src/main/java/com/lingo_leap/repository/TokenRepository.java
@@ -12,7 +12,7 @@ public interface TokenRepository extends JpaRepository<Token, Integer> {
 
 
     @Query("select t from Token t inner join User u on t.user.id = u.id where t.user.id = :userId and t.loggedOut = false")
-    List<Token> findAllAccessTokensByUser(Integer userId);
+    List<Token> findAllAccessTokensByUser(Long userId);
 
     Optional<Token> findByAccessToken(String token);
